feat: emit one Handle impl and the value-definition wiring from the subscriber attribute
The attribute becomes sugar over the manual path: the expansion generates the
handler's impl Handle plus a Declared whose settings wrap the same sealed
definition the subscriber(..) chain builds, carrying probe-captured metadata
as data. A handler with Out parameters keeps the unit struct at the include
site and instantiates the arena definition in BindSlots::bind, so every
include-site chain (.publisher(..), .out(marker, ..), .build()) keeps
compiling unchanged.

The combinations the unified rails cannot express without losing a feature
keep the definition-trait emission point-wise: Seek(..) parameters (a
different broker capability resolves them), an Out parameter's declared
message set (the arena entry carries no per-parameter dictionary), the raw
batch shape (rebinding would add a per-page allocation), and a batch reading
Headers<Vec<_>> contracts (the page signature carries no per-element header
source).

The manual state-agnostic bodies respell to the defaulted state axis: with
the state quantified at the adapters, the constructor no longer infers a
generic body's state parameter from the mount.
