+++ tests/ui/publish_outside_dictionary.stderr
@@ -20,12 +20,12 @@ help: the trait `PublishedThrough<Events>` is implemented for `Progress`
    |
 22 | #[derive(OutSlot)]
    |          ^^^^^^^
-note: required by a bound in `TypedSlot::<P, Body, M, EncodeCodec>::message`
-  --> src/runtime/slot.rs
+note: required by a bound in `Slot::<M, W, E>::message`
+  --> src/runtime/handle/outs.rs
    |
-   |     pub fn message<'a, T, Index>(
+   |     pub fn message<'a, T>(
    |            ------- required by a bound in this associated function
 ...
    |         T: OutgoingDestination + PublishedThrough<M>,
-   |                                  ^^^^^^^^^^^^^^^^^^^ required by this bound in `TypedSlot::<P, Body, M, EncodeCodec>::message`
+   |                                  ^^^^^^^^^^^^^^^^^^^ required by this bound in `Slot::<M, W, E>::message`
    = note: this error originates in the derive macro `OutSlot` (in Nightly builds, run with -Z macro-backtrace for more info)
