+++ tests/ui/subscriber_out_capability_mismatch.stderr
@@ -29,14 +29,16 @@ help: the following other types implement trait `RequestReply`
    |
    |   impl RequestReply for MemoryRequester {
    |   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `MemoryRequester`
-   = note: required for `SlotPublisher<MemoryPublisher, DefaultSlot>` to implement `RequestReply`
-note: required for `__RsOutDef_forward<SlotPublisher<MemoryPublisher, DefaultSlot>, JsonCodec>` to implement `InjectDef`
+note: required for `forward` to implement `Handle<Order, (), Outs<(Slot<DefaultSlot, MemoryPublisher, JsonCodec>,)>>`
   --> tests/ui/subscriber_out_capability_mismatch.rs:14:1
    |
 14 | #[subscriber("orders")]
    | ^^^^^^^^^^^^^^^^^^^^^^^
 15 | async fn forward(order: &Order, Out(_out): Out<impl RequestReply>) -> HandlerOutcome {
-   |                                                     ------------ unsatisfied trait bound introduced here
+   |          ^^^^^^^                                    ------------ unsatisfied trait bound introduced here
+   = note: required for `Sealed<HandleValue<Solo<Order>, (), Outs<(Slot<DefaultSlot, MemoryPublisher, JsonCodec>,)>, (), forward, ruststream::runtime::Probed>>` to implement `InjectCall<()>`
+   = note: 1 redundant requirement hidden
+   = note: required for `SubscriberBuilder<Sealed<HandleValue<Solo<Order>, (), Outs<(Slot<DefaultSlot, MemoryPublisher, JsonCodec>,)>, (), forward, ruststream::runtime::Probed>>, Name, (Open, Open, Open)>` to implement `InjectCall<()>`
    = note: required for `(ruststream::runtime::WithSource<MemoryPublish>,)` to implement `ruststream::runtime::SlotCommit<runtime::router::mount::InjectMount, MemoryBroker, Identity, (), (), PublishIdentity, SubscriberBuilder<forward, Name, (Open, Open, Open)>>`
 note: required by a bound in `IncludeSlots::<'_, Mount, B, Layers, C, State, Pipeline, Def, (ruststream::runtime::MissingSlot<M>,)>::publisher`
   --> src/runtime/app/include/slot_builder.rs
