+++ tests/ui/subscriber_publish_not_str.stderr
@@ -2,7 +2,10 @@ error[E0308]: mismatched types
  --> tests/ui/subscriber_publish_not_str.rs:5:32
   |
 5 | #[subscriber("orders", publish(123))]
-  |                                ^^^ expected `&str`, found integer
+  | -------------------------------^^^---
+  | |                              |
+  | |                              expected `&str`, found integer
+  | expected due to this
 
 error[E0308]: mismatched types
  --> tests/ui/subscriber_publish_not_str.rs:5:32
@@ -11,4 +14,10 @@ error[E0308]: mismatched types
   | -------------------------------^^^---
   | |                              |
   | |                              expected `&str`, found integer
-  | expected due to this
+  | arguments to this function are incorrect
+  |
+note: function defined here
+ --> src/runtime/handle/value.rs
+  |
+  | pub fn probed_reply_def<A, R, O, C, H, Route, Attach>(
+  |        ^^^^^^^^^^^^^^^^
